requirements.txt
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.4
packaging==24.0
picamera==1.12
pillow==10.3.0
pip==24.0
rpi-kms==0.1a1
rpi-libcamera==0.1a4
setuptools==41.2.0
wheel==0.43.0
XlsxWriter==3.2.0
